Adding
Can add a file to a folder in the tree. Need to figure out how to add a folder. Deleting works, however, if the file you are trying to delete is the oly child of a parent folder, the file deletes and tthe parent folder turns into a file. No idea why thats the case.