Add files via upload
Mital and I worked on this together as out IntelliJ was having an issue that we couldnt fix. We ran this on an old build to ensure it was working and then pushed it to the new folder. We are still working on the other parts, but it is taking alot longer not being able to ensure working compatibility with everyone else.