Updates
Open File and Save Project are not functional.

Before they would throw errors, now they are able to correctly display a text file (or any other viewable file) in the CodeEditor windows, they allow edits to be made, and when all is done and you click Save Project, the changes are saved to the file on your machine.